requirements.txt
# 本地打包使用国内镜像源加速下载， 默认不使用，影响github actions 
# --index-url https://pypi.tuna.tsinghua.edu.cn/simple

# 基础科学计算和数据处理库
numpy==2.1.2
pandas==2.2.2

# 数据获取和分析库
akshare

# 网络和API请求
requests==2.32.3
python-dotenv==1.0.1
flask==3.1.0
flask-jwt-extended==4.7.1

# 日志和系统工具
loguru==0.7.2

# 数据库和ORM
flask-sqlalchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10